✨ ♻️ move lib::main into main.rs and add tracing config
