+++ .env.example
@@ -5,7 +5,6 @@ PROFILE=local
 BACKEND_GRAPHQL_API_KEY=myapikey
 
 # Logs
-ANSI_LOGS=true
 RUST_LOG=info
 
 # Hasura
