+++ github-proxy/src/main.rs
@@ -1,17 +1,27 @@
 use anyhow::Result;
 use dotenv::dotenv;
-use infrastructure::tracing::Tracer;
+use infrastructure::{
+	config,
+	tracing::{self, Tracer},
+};
 
 #[macro_use]
 extern crate rocket;
 
 mod presentation;
 use presentation::http;
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+struct Config {
+	tracer: tracing::Config,
+}
 
 #[rocket::main]
 async fn main() -> Result<()> {
 	dotenv().ok();
-	let _tracer = Tracer::init("api")?;
+	let config: Config = config::load("github-proxy/app.yaml")?;
+	let _tracer = Tracer::init(&config.tracer, "github-proxy")?;
 
 	http::serve().await?;
